refactor: move `middleware-aborter` into `middlewares`
