--- packages/middleware-aborter/CHANGELOG.md
@@ -1,13 +1,0 @@
-## v2.0.0
-
--   🔧 Refactor AbortEventHandler
-
-## v1.4.1
-
--   🐛 Fix error which will not trigger `onAbort` in `timeout` ([#160](https://github.com/fanhaoyuan/fatcher/pull/160))
-
-## v1.4.0
-
--   🚀 Add `isAbortError` helpers. ([#157](https://github.com/fanhaoyuan/fatcher/pull/157))
--   🚀 Add `concurrency` prop for aborting concurrency request ([#158](https://github.com/fanhaoyuan/fatcher/pull/158))
--   🐛 Remove unexpected warning when `timeout = 0` ([#159](https://github.com/fanhaoyuan/fatcher/pull/159))
