--- packages/middleware-aborter/package.json
@@ -1,28 +1,0 @@
-{
-    "name": "@fatcherjs/middleware-aborter",
-    "version": "2.0.0",
-    "main": "dist/aborter.js",
-    "module": "dist/aborter.esm.js",
-    "browser": "dist/aborter.min.js",
-    "typings": "dist/aborter.d.ts",
-    "scripts": {
-        "dev": "rimraf dist && rollup -c rollup.config.ts -w",
-        "build": "rimraf dist && rollup -c rollup.config.ts",
-        "deploy": "pnpm run build && pnpm publish --no-git-check"
-    },
-    "license": "MIT",
-    "files": [
-        "dist"
-    ],
-    "homepage": "https://github.com/fanhaoyuan/fatcher/tree/master/packages/middleware-aborter",
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/fanhaoyuan/fatcher.git"
-    },
-    "devDependencies": {
-        "fatcher": "workspace:^"
-    },
-    "peerDependencies": {
-        "fatcher": "^2.0.0"
-    }
-}
