--- packages/middleware-aborter/tsconfig.json
@@ -1,5 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "include": ["src/**/*", "__tests__/**/*"],
-    "exclude": ["node_modules", "dist"]
-}
